[NEST-FE-26]Extract ImageUpload as a smaller component from CreatePost component
